Sage trac #9900: Be more careful with conjugate(f(x)) -> f(conjugate(x)).
Based on the patch by Richard Kreckel for GiNaC, which can be found in the URL:

http://www.ginac.de/ginac.git?p=ginac.git;a=commitdiff_plain;h=ff8b400eb500618644231ed9e6f199c3b0b25135;hp=1a5127603d93125b75ee51c94200faf0b586ef32

Be more careful with conjugate(f(x)) -> f(conjugate(x)).

That identity is correct for holomorphic functions, but we have to be
careful with some functions not being holomorphic everywhere. On branch
cuts, it is wrong. For a discussion, see:
<http://www.ginac.de/pipermail/ginac-list/2010-April/001601.html>.

This patch changes the default behavior of user-defined functions. From
now on, a user-defined conjugate_func must be registered, in order to
evaluate conjugate(f(x)) -> f(conjugate(x)). This patch also adds such
functions for most initially known functions.