Sage #12289: change argument of evalf to an arbitrary dict
evalf() method of symbolic functions pass around a Python object while
traversing the expression tree. When calling a numeric evalution method
defined in Python, this argument is put in a dictionary with the keyword
'parent'.

This patch treats the Python object passed around in evalf() methods as a
dictionary, allowing arbitrary parameters to be stored in it..