Fix unarchiving of symbolic functions defined at runtime without any custom methods.
Since the python_func member of function is a bitmask indicating which custom
methods are defined in Python, symbolic functions defined at runtime are
archived as regular GiNaC functions defined in C++. Unarchiving these later is
not possible, as GiNaC does not expect any other functions than those already
in the registry.

Adding a Python callback to create a new symbolic function with the given and
number of arguments if one is not found in the registry solves the unarchiving
problem while keeping the archive/pickle format the same.