fixes #12 : infinite loop when normalizing an add with content in GF(2)
If the content of an add expression is in GF(2^k), normalizing this add within
a mul::eval() leads to infinite loops. The normalization is done by
multiplying by -c where c is the content. When characteristic is 2 and c is 1,
-c is also 1. So no normalization occurs, causing the eval() function to run
into this case again.

This patch introduces a check if the characteristic is 2 when setting the
normalization factor -c. In this case, the factor -1 in ZZ is used instead to
fix the sign of the leading term that is triggering the normalization in the
first place.

This was #13609 on Sage trac.