Fix compliation error with GMP long long limbs.
Depending how GMP is configured the macro from flint, slong, which actually
refers to GMP's mp_limb_signed_t, may be a long long instead of plain long.

numeric doesn't have a constructor for long long, leading to an ambiguous
overload error on the constructor.  In this case it should make more sense
to just use normal int and long types for these loop variables as it
makes sense to, and only cast to the "slong" type when passing to the
flint interface.